executor: move current-thread into crate
The `CurrentThread` executor is exposed using a feature flag.

Refs: #1264
