+++ azure-pipelines.yml
@@ -78,6 +78,16 @@ jobs:
         - async-traits
       tokio-test: []
 
+# Test compilation failure
+- template: ci/azure-test-stable.yml
+  parameters:
+    name: test_features
+    displayName: Test feature flags
+    rust: $(nightly)
+    crates:
+      ui-tests:
+        - executor-without-current-thread
+
 # - template: ci/azure-cargo-check.yml
 #   parameters:
 #     name: features
