--- tokio-current-thread/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "tokio-current-thread"
-# When releasing to crates.io:
-# - Remove path dependencies
-# - Update html_root_url.
-# - Update doc url
-#   - Cargo.toml
-# - Update CHANGELOG.md.
-# - Create "v0.2.x" git tag.
-version = "0.2.0-alpha.1"
-edition = "2018"
-authors = ["Tokio Contributors <team@tokio.rs>"]
-license = "MIT"
-repository = "https://github.com/tokio-rs/tokio"
-homepage = "https://tokio.rs"
-documentation = "https://docs.rs/tokio-current-thread/0.2.0-alpha.1/tokio_current_thread"
-description = """
-Single threaded executor which manage many tasks concurrently on the current thread.
-"""
-keywords = ["futures", "tokio"]
-categories = ["concurrency", "asynchronous"]
-
-[dependencies]
-tokio-executor = { version = "=0.2.0-alpha.1", path = "../tokio-executor" }
-crossbeam-channel = "0.3.8"
-
-[dev-dependencies]
-tokio-sync = { version = "=0.2.0-alpha.1", path = "../tokio-sync" }
