--- tokio-current-thread/README.md
@@ -1,13 +1,0 @@
-# tokio-current-thread
-
-Single threaded executor for Tokio.
-
-## License
-
-This project is licensed under the [MIT license](LICENSE).
-
-### Contribution
-
-Unless you explicitly state otherwise, any contribution intentionally submitted
-for inclusion in Tokio by you, shall be licensed as MIT, without any additional
-terms or conditions.
