+++ tokio-executor/src/lib.rs
@@ -64,6 +64,9 @@ mod global;
 pub mod park;
 mod typed;
 
+#[cfg(feature = "current-thread")]
+pub mod current_thread;
+
 pub use crate::enter::{enter, exit, Enter, EnterError};
 pub use crate::error::SpawnError;
 pub use crate::executor::Executor;
