+++ tokio-timer/Cargo.toml
@@ -37,7 +37,6 @@ slab = "0.4.1"
 
 [dev-dependencies]
 tokio = { version = "=0.2.0-alpha.1", path = "../tokio" }
-tokio-current-thread = { version = "=0.2.0-alpha.1", path = "../tokio-current-thread" }
 tokio-sync = { version = "=0.2.0-alpha.1", path = "../tokio-sync", features = ["async-traits"] }
 tokio-test = { version = "=0.2.0-alpha.1", path = "../tokio-test" }
