+++ tokio/Cargo.toml
@@ -45,8 +45,7 @@ rt-full = [
   "reactor",
   "sync",
   "timer",
-  "tokio-current-thread",
-  "tokio-executor",
+  "tokio-executor/current-thread",
   "tokio-macros",
   "tokio-threadpool",
   "tracing-core",
@@ -66,7 +65,6 @@ futures-util-preview = { version = "=0.3.0-alpha.18", features = ["sink"] }
 bytes = { version = "0.4", optional = true }
 num_cpus = { version = "1.8.0", optional = true }
 tokio-codec = { version = "=0.2.0-alpha.1", optional = true, path = "../tokio-codec" }
-tokio-current-thread = { version = "=0.2.0-alpha.1", optional = true, path = "../tokio-current-thread" }
 tokio-fs = { version = "=0.2.0-alpha.1", optional = true, path = "../tokio-fs" }
 tokio-io = { version = "=0.2.0-alpha.1", optional = true, features = ["util"], path = "../tokio-io" }
 tokio-executor = { version = "=0.2.0-alpha.1", optional = true, path = "../tokio-executor" }
