+++ ui-tests/src/lib.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "tokio-executor")]
+pub use tokio_executor;
