+++ ui-tests/tests/ui/executor_without_current_thread.rs
@@ -1,0 +1,3 @@
+use ui_tests::tokio_executor::current_thread;
+
+fn main() {}
