+++ ui-tests/tests/ui/executor_without_current_thread.stderr
@@ -1,0 +1,7 @@
+error[E0432]: unresolved import `ui_tests::tokio_executor::current_thread`
+ --> $DIR/executor_without_current_thread.rs:1:5
+  |
+1 | use ui_tests::tokio_executor::current_thread;
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ no `current_thread` in `tokio_executor`
+
+For more information about this error, try `rustc --explain E0432`.
